+++ CHANGELOG.md
@@ -46,6 +46,7 @@ xx
 ### majorization.v
 
 #### Added
+
 - definitions `majorize`, `weak_majorize`
 - lemmas `sort_v_sum`, `weak_majorize_ltP`, `weak_majorize_leP`, `majorize_ltP`, `majorize_leP`, `majorizeW`, `majorize_refl`, `weak_majorize_refl`, `majorize_trans`, `weak_majorize_trans`, `weak_majorize_anti`, `majorize_anti`, `sort_v_eq`
 - definitions `doubly_substochastic`, `doubly_stochastic`, `elem_lemx`
@@ -62,6 +63,17 @@ xx
 - lemmas `xlnx_sum_fin`, `xlnx_sum`, `is_derive1_xlnx`, `is_derive12_xlnx`, `continuous_xlnx`, `ln1x_le`, `xlnx_cvg`, `convex_xlnx`, `xlnx_average_sum_ord`, `xlnx_average_sum`
 - lemmas `powR_weak_majorize`, `exp_ln_weak_majorize`, `ln_weak_majorize`, `ln_prod`, `prod_sum_weak_majorize_ln`, `prod_sum_weak_majorize_gt0`, `prod_sum_weak_majorize`, `entropy_majority`, `weak_majorize_sum`, `majority_entropy_le`
 - lemmas `svd_fRE`, `svd_fE`, `svd_fR_nincr`, `svd_fR_ge0`, `svd_fR_nneg`, `svd_fR_gt0`, `svd_fR_eq0`, `svd_fR_pos`. `svd_fR_prodM`, `weak_majorize_svd_fR`, `svd_fR_sumM`, `svd_f_sumM`, `svd_fR_powM`, `svd_fR_fM`
+
+### mcaextra.v
+
+#### Added
+
+- definition `directc`
+- lemmas `directc_norm`, `norm_directcE`
+
+#### Removed
+
+- lemmas `split2c`, `split2l`, `split2`
 
 ### mcextra.v